:muscle: separate modules types and error classes
