+++ mod.ts
@@ -23,11 +23,5 @@
  * @module
  */
 
-export { BatchError } from "./denops.ts";
-export type {
-  Context,
-  Denops,
-  Dispatcher,
-  Entrypoint,
-  Meta,
-} from "./denops.ts";
+export * from "./errors.ts";
+export type * from "./types.ts";
