+++ types.ts
@@ -1,4 +1,12 @@
 /**
+ * This module provides type interfaces for [denops.vim].
+ *
+ * [denops.vim]: https://github.com/vim-denops/denops.vim
+ *
+ * @module
+ */
+
+/**
  * API dispatcher
  */
 export type Dispatcher = {
@@ -32,27 +40,6 @@ export interface Meta {
    * Host platform name.
    */
   readonly platform: "windows" | "mac" | "linux";
-}
-
-/**
- * Batch error raised when one of the functions fails during batch process.
- */
-export class BatchError extends Error {
-  /**
-   * A result list that is successfully completed prior to the error.
-   */
-  readonly results: unknown[];
-
-  constructor(message: string, results: unknown[]) {
-    super(message);
-
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, BatchError);
-    }
-
-    this.name = this.constructor.name;
-    this.results = results;
-  }
 }
 
 /**
